refactor!: remove NO_MARGINS constant and have 0 px margins as default
For simplicity, the default is to not have any margins between the chart rendering and the canvas edge.

BREAKING CHANGE: the NO_MARGINS constant was removed and by default there are no margins. To migrate, remove any use of NO_MARGINS. To mimic the old behaviour,  set the margins to 2 px where none have been set before.